refactor(linter): share per-file shell resolution
