+++ crates/shuck-linter/src/lib.rs
@@ -104,12 +104,13 @@ pub use rule_selector::{RuleSelector, SelectorParseError};
 pub use rule_set::RuleSet;
 #[allow(unused_imports)]
 pub(crate) use rules::common::word::conditional_binary_op_is_string_match;
-/// Linter configuration and per-file ignore types.
+/// Linter configuration and per-file selection types.
 pub use settings::{
     AmbientShellOptions, C001RuleOptions, C063RuleOptions, C158RuleOptions, C159RuleOptions,
-    C160RuleOptions, C161RuleOptions, CompiledPerFileIgnoreList, LinterRuleOptions, LinterSettings,
-    PerFileIgnore, S078RuleOptions, S079RuleOptions, S080RuleOptions, S081RuleOptions,
-    S082RuleOptions, S083FunctionDocRequirement, S083RuleOptions, S084RuleOptions, S085RuleOptions,
+    C160RuleOptions, C161RuleOptions, CompiledPerFileIgnoreList, CompiledPerFileShellList,
+    LinterRuleOptions, LinterSettings, PerFileIgnore, PerFileShell, S078RuleOptions,
+    S079RuleOptions, S080RuleOptions, S081RuleOptions, S082RuleOptions, S083FunctionDocRequirement,
+    S083RuleOptions, S084RuleOptions, S085RuleOptions,
 };
 /// Shell dialect selection used by the linter.
 pub use shell::ShellDialect;
