+++ crates/shuck-server/Cargo.toml
@@ -14,7 +14,6 @@ fuzzing = []
 [dependencies]
 anyhow = { workspace = true }
 crossbeam = { workspace = true }
-globset = { workspace = true }
 lsp-server = { workspace = true }
 lsp-types = { workspace = true }
 rustc-hash = { workspace = true }
